refactor: circuits (#80)
