+++ apps/next/hooks/use-promote-post.ts
@@ -1,5 +1,5 @@
 import { api } from '@/lib/api'
-import { generateProof, ProofType } from '@anon/utils/src/proofs'
+import { generateProof, ProofType } from '@anon/utils/src/proofs/generate'
 import { useState } from 'react'
 import { hashMessage } from 'viem'
 import { useAccount, useSignMessage } from 'wagmi'
@@ -74,14 +74,14 @@ export const usePromotePost = (tokenAddress: string) => {
       if (process.env.NEXT_PUBLIC_DISABLE_QUEUE) {
         await api.promotePost(
           Array.from(proof.proof),
-          proof.publicInputs.map((i) => Array.from(i)),
+          proof.publicInputs.map((input) => Array.from(input)),
           { asReply }
         )
       } else {
         await api.submitAction(
           ProofType.PROMOTE_POST,
           Array.from(proof.proof),
-          proof.publicInputs.map((i) => Array.from(i)),
+          proof.publicInputs.map((input) => Array.from(input)),
           { asReply }
         )
       }
