+++ apps/next/next.config.mjs
@@ -1,4 +1,51 @@
+import webpack from 'webpack'
+
 /** @type {import('next').NextConfig} */
-const nextConfig = {}
+const nextConfig = {
+  experimental: {
+    outputFileTracingIncludes: {
+      '/': [
+        './node_modules/@aztec/bb.js/dest/node/barretenberg_wasm/**/*',
+        './node_modules/@aztec/bb.js/dest/node/barretenberg_wasm/barretenberg_wasm_thread/factory/node/thread.worker.js',
+      ],
+      '//': [
+        './node_modules/@aztec/bb.js/dest/node/barretenberg_wasm/**/*',
+        './node_modules/@aztec/bb.js/dest/node/barretenberg_wasm/barretenberg_wasm_thread/factory/node/thread.worker.js',
+      ],
+    },
+  },
+  reactStrictMode: false,
+  webpack: (config) => {
+    config.experiments = {
+      asyncWebAssembly: true,
+      syncWebAssembly: true,
+      layers: true,
+    }
+    config.externals.push('pino-pretty', 'lokijs', 'encoding')
+    config.plugins.push(
+      new webpack.NormalModuleReplacementPlugin(/^node:/, (resource) => {
+        resource.request = resource.request.replace(/^node:/, '')
+      })
+    )
+    return config
+  },
+  async headers() {
+    return [
+      {
+        source: '/:path*',
+        headers: [
+          {
+            key: 'Cross-Origin-Embedder-Policy',
+            value: 'require-corp',
+          },
+          {
+            key: 'Cross-Origin-Opener-Policy',
+            value: 'same-origin',
+          },
+        ],
+      },
+    ]
+  },
+}
 
 export default nextConfig
