+++ apps/next/package.json
@@ -41,7 +41,8 @@
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
     "viem": "2.x",
-    "wagmi": "^2.12.33"
+    "wagmi": "^2.12.33",
+    "webpack": "^5.96.1"
   },
   "devDependencies": {
     "@types/circomlibjs": "^0",
