+++ packages/api/package.json
@@ -4,10 +4,7 @@
   "private": true,
   "types": "./src",
   "main": "src/index.ts",
-  "files": [
-    "types",
-    "dist"
-  ],
+  "files": ["types", "dist"],
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "bun run --watch src/index.ts",
@@ -17,9 +14,9 @@
     "@anon/circuits": "*",
     "@anon/db": "*",
     "@anon/utils": "*",
-    "@elysiajs/cors": "^1.1.1",
     "@noir-lang/backend_barretenberg": "^0.17.0",
     "@noir-lang/noir_js": "^0.17.0",
+    "@elysiajs/cors": "^1.1.1",
     "elysia": "^1.1.25",
     "ioredis": "^5.4.1",
     "logestic": "^1.2.4",
