+++ packages/api/src/index.ts
@@ -5,7 +5,8 @@ import { merkleTreeRoutes } from './routes/merkle-tree'
 import { getPostRoutes } from './routes/post'
 import { uploadRoutes } from './routes/upload'
 import { neynar } from './services/neynar'
-import { getProvingBackend, ProofType } from '@anon/utils/src/proofs'
+import { ProofType } from '@anon/utils/src/proofs/generate'
+import { getProvingBackend } from '@anon/utils/src/proofs/verify'
 ;(async () => {
   const [createPostBackend, submitHashBackend] = await Promise.all([
     getProvingBackend(ProofType.CREATE_POST),
