+++ packages/api/src/routes/merkle-tree.ts
@@ -1,7 +1,7 @@
 import { createElysia } from '../utils'
 import { t } from 'elysia'
 import { buildHoldersTree, getTree, setTree } from '@anon/utils/src/merkle-tree'
-import { ProofType } from '@anon/utils/src/proofs'
+import { ProofType } from '@anon/utils/src/proofs/generate'
 import { TOKEN_CONFIG } from '@anon/utils/src/config'
 
 export const merkleTreeRoutes = createElysia({ prefix: '/merkle-tree' }).post(
