+++ packages/api/src/utils.ts
@@ -3,4 +3,9 @@ import { Elysia } from 'elysia'
 import { Logestic } from 'logestic'
 
 export const createElysia = (config?: ConstructorParameters<typeof Elysia>[0]) =>
-  new Elysia(config).use(cors()).use(Logestic.preset('common'))
+  new Elysia(config)
+    .use(cors())
+    .use(Logestic.preset('common'))
+    .onError(({ code, error }) => {
+      console.error(error)
+    })
