+++ packages/circuits/mimc/Nargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "mimc"
+type = "lib"
+authors = [""]
+compiler_version = ">=0.35.0"
+
+[dependencies]
