+++ packages/circuits/submit-hash/src/main.nr
@@ -1,5 +1,5 @@
-use dep::std;
 use dep::ecrecover;
+use dep::std;
 
 fn main(
     // Token Balance
