+++ packages/utils/package.json
@@ -4,10 +4,7 @@
   "private": true,
   "types": "./src",
   "main": "src/index.ts",
-  "files": [
-    "types",
-    "dist"
-  ],
+  "files": ["types", "dist"],
   "scripts": {
     "update-merkle-tree": "bun run src/update-merkle-tree.ts"
   },
